refactor(webview): extract mixins, reduce to 970 lines (Iteration #132)
- Create WebViewStateMixin, WebViewCommandsMixin, WebViewChannelsMixin
- Create WebViewFactoryMixin, WebViewBridgeMixin
- Remove duplicate methods from webview.py
- webview.py: 1541 -> 970 lines (target: <1000)
