+++ python/auroraview/core/webview.py
@@ -24,6 +24,11 @@ from auroraview.core.mixins import (
     WebViewLifecycleMixin,
     WebViewTelemetryMixin,
     WebViewWindowMixin,
+    WebViewStateMixin,
+    WebViewCommandsMixin,
+    WebViewChannelsMixin,
+    WebViewFactoryMixin,
+    WebViewBridgeMixin,
 )
 
 if TYPE_CHECKING:
@@ -58,6 +63,11 @@ class WebView(
     WebViewApiMixin,
     WebViewDOMMixin,
     WebViewTelemetryMixin,
+    WebViewStateMixin,
+    WebViewCommandsMixin,
+    WebViewChannelsMixin,
+    WebViewFactoryMixin,
+    WebViewBridgeMixin,
 ):
     """High-level WebView class with enhanced Python API.
 
@@ -555,458 +565,11 @@ class WebView(
         # Initialize auto-telemetry (after WindowManager registration)
         self._init_telemetry()
 
-    @property
-    def state(self) -> "State":
-        """Get the shared state container for Python ↔ JavaScript sync.
 
-        Returns:
-            State container with dict-like interface
 
-        Example:
-            >>> webview.state["user"] = {"name": "Alice"}
-            >>> webview.state["theme"] = "dark"
-            >>>
-            >>> @webview.state.on_change
-            >>> def handle_change(key, value, source):
-            ...     print(f"{key} = {value} from {source}")
-        """
-        if self._state is None:
-            from .state import State
 
-            self._state = State(self)
-        return self._state
 
-    @property
-    def commands(self) -> "CommandRegistry":
-        """Get the command registry for Python ↔ JavaScript RPC.
 
-        Returns:
-            CommandRegistry instance
-
-        Example:
-            >>> @webview.commands.register
-            >>> def greet(name: str) -> str:
-            ...     return f"Hello, {name}!"
-        """
-        if self._commands is None:
-            from .commands import CommandRegistry
-
-            self._commands = CommandRegistry(self)
-        return self._commands
-
-    def command(self, func_or_name=None):
-        """Decorator to register a command callable from JavaScript.
-
-        This is a convenience shortcut for `webview.commands.register`.
-
-        Args:
-            func_or_name: Function to register or custom command name
-
-        Returns:
-            Decorated function
-
-        Example:
-            >>> @webview.command
-            >>> def greet(name: str) -> str:
-            ...     return f"Hello, {name}!"
-            >>>
-            >>> @webview.command("add_numbers")
-            >>> def add(x: int, y: int) -> int:
-            ...     return x + y
-            >>>
-            >>> # In JavaScript:
-            >>> # const msg = await auroraview.invoke("greet", {name: "World"});
-            >>> # const sum = await auroraview.invoke("add_numbers", {x: 1, y: 2});
-        """
-        return self.commands.register(func_or_name)
-
-    @property
-    def channels(self) -> "ChannelManager":
-        """Get the channel manager for streaming data.
-
-        Returns:
-            ChannelManager instance
-
-        Example:
-            >>> channel = webview.channels.create()
-            >>> channel.send({"progress": 50})
-            >>> channel.send({"progress": 100})
-            >>> channel.close()
-        """
-        if self._channels is None:
-            from .channel import ChannelManager
-
-            self._channels = ChannelManager(self)
-        return self._channels
-
-    def create_channel(self, channel_id: Optional[str] = None) -> "Channel":
-        """Create a new streaming channel.
-
-        This is a convenience shortcut for `webview.channels.create()`.
-
-        Args:
-            channel_id: Optional custom channel ID
-
-        Returns:
-            New Channel instance
-
-        Example:
-            >>> with webview.create_channel() as channel:
-            ...     for chunk in read_large_file():
-            ...         channel.send(chunk)
-        """
-        return self.channels.create(channel_id)
-
-    @classmethod
-    def create(
-        cls,
-        title: str = "AuroraView",
-        *,
-        # Content
-        url: Optional[str] = None,
-        html: Optional[str] = None,
-        # Window properties
-        width: int = 800,
-        height: int = 600,
-        resizable: bool = True,
-        frame: Optional[bool] = None,
-        always_on_top: bool = False,
-        transparent: bool = False,
-        background_color: Optional[str] = None,
-        # DCC integration
-        parent: Optional[int] = None,
-        mode: Literal["auto", "owner", "child", "container"] = "auto",
-        # Bridge integration
-        bridge: Union["Bridge", bool, None] = None,  # type: ignore
-        # Development options
-        debug: bool = True,
-        context_menu: bool = True,
-        # Custom protocol
-        asset_root: Optional[str] = None,
-        data_directory: Optional[str] = None,
-        allow_file_protocol: bool = False,
-        # Automation
-        auto_show: bool = False,
-        auto_timer: bool = True,
-        # Singleton control
-        singleton: Optional[str] = None,
-        # IPC performance tuning
-        ipc_batch_size: int = 0,
-        # Custom icon
-        icon: Optional[str] = None,
-        # Window style
-        tool_window: bool = False,
-        undecorated_shadow: bool = False,
-        # New window handling
-        allow_new_window: bool = False,
-        new_window_mode: Optional[str] = None,
-        # Remote debugging
-        remote_debugging_port: Optional[int] = None,
-    ) -> "WebView":
-        """Create WebView instance (recommended way).
-
-        Args:
-            title: Window title
-            url: URL to load
-            html: HTML content to load
-            width: Window width in pixels
-            height: Window height in pixels
-            resizable: Make window resizable
-            frame: Show window frame (title bar, borders). If None, uses sensible defaults.
-
-            always_on_top: Keep window always on top of other windows (default: False)
-            parent: Parent window handle for DCC embedding
-            mode: Embedding mode
-                - "auto": Auto-select (recommended)
-                - "owner": Owner mode (cross-thread safe)
-                - "child": Child window mode (same-thread)
-            bridge: Bridge for DCC/Web integration
-                - Bridge instance: Use provided bridge
-                - True: Auto-create bridge (port 9001)
-                - None: No bridge (default)
-            debug: Enable developer tools
-            context_menu: Enable native context menu (default: True)
-            asset_root: Root directory for auroraview:// protocol
-            data_directory: User data directory for WebView (cookies, cache, localStorage).
-                If None, uses system default. Set to isolate data per app/user profile.
-            allow_file_protocol: Enable file:// protocol support (default: False)
-                WARNING: Enabling this bypasses WebView's default security restrictions
-            auto_show: Automatically show after creation
-            auto_timer: Auto-start event timer for embedded mode (recommended)
-            singleton: Singleton key. If provided, only one instance with this key
-                      can exist at a time. Calling create() again with the same key
-                      returns the existing instance.
-
-        Returns:
-            WebView instance
-
-        Examples:
-            >>> # Standalone window
-            >>> webview = WebView.create("My App", url="http://localhost:3000")
-            >>> webview.show()
-
-            >>> # DCC embedding (Maya)
-            >>> webview = WebView.create("Maya Tool", parent=maya_hwnd)
-            >>> webview.show()
-
-            >>> # With Bridge integration
-            >>> webview = WebView.create("Photoshop Tool", bridge=True)
-            >>> @webview.bridge.on('layer_created')
-            >>> async def handle_layer(data, client):
-            ...     return {"status": "ok"}
-            >>> webview.show()
-
-            >>> # Auto-show
-            >>> webview = WebView.create("App", auto_show=True)
-
-            >>> # Singleton mode - only one instance allowed
-
-        Note:
-            For Qt-based DCC applications (Maya, Houdini, Nuke), consider using
-            QtWebView instead for automatic event processing and better integration:
-
-            >>> from auroraview import QtWebView
-            >>> webview = QtWebView(parent=maya_main_window(), title="My Tool")
-            >>> webview.load_url("http://localhost:3000")
-            >>> webview.show()  # Automatic event processing!
-            >>> webview1 = WebView.create("Tool", singleton="my_tool")
-            >>> webview2 = WebView.create("Tool", singleton="my_tool")  # Returns webview1
-            >>> assert webview1 is webview2
-        """
-        # Check singleton registry
-        if singleton is not None:
-            if singleton in cls._singleton_registry:
-                existing = cls._singleton_registry[singleton]
-                logger.info(f"Returning existing singleton instance: '{singleton}'")
-                return existing
-            logger.info(f"Creating new singleton instance: '{singleton}'")
-        # Detect mode
-        is_embedded = parent is not None
-
-        # Auto-select mode
-        if mode == "auto":
-            actual_mode = "owner" if is_embedded else None
-            if is_embedded:
-                logger.info(f"[AUTO-DETECT] parent={parent} detected, auto-selecting mode='owner'")
-        else:
-            actual_mode = mode if is_embedded else None
-            if is_embedded:
-                logger.info(f"[MANUAL] Using user-specified mode='{mode}'")
-
-        logger.info(f"[MODE] Final mode: {actual_mode} (embedded={is_embedded})")
-
-        # Create instance
-        # For embedded mode, always set auto_show=False to let Qt control visibility
-        # For standalone mode, use the user-provided auto_show value
-        rust_auto_show = False if is_embedded else auto_show
-        instance = cls(
-            title=title,
-            width=width,
-            height=height,
-            url=url,
-            html=html,
-            resizable=resizable,
-            frame=frame,
-            always_on_top=always_on_top,
-            transparent=transparent,
-            background_color=background_color,
-            parent=parent,
-            mode=actual_mode,
-            debug=debug,
-            context_menu=context_menu,
-            bridge=bridge,
-            asset_root=asset_root,
-            data_directory=data_directory,
-            allow_file_protocol=allow_file_protocol,
-            auto_show=rust_auto_show,  # Pass to Rust layer
-            ipc_batch_size=ipc_batch_size,  # Max messages per tick (0=unlimited)
-            icon=icon,  # Custom window icon path
-            tool_window=tool_window,  # Tool window style
-            undecorated_shadow=undecorated_shadow,  # Shadow for frameless windows
-            allow_new_window=allow_new_window,  # Allow window.open()
-            new_window_mode=new_window_mode,  # New window behavior
-            remote_debugging_port=remote_debugging_port,  # CDP debugging port
-        )
-
-        # Auto timer (embedded mode)
-        if is_embedded and auto_timer:
-            try:
-                from auroraview.utils.event_timer import EventTimer
-
-                instance._auto_timer = EventTimer(instance, interval_ms=16)
-                instance._auto_timer.on_close(lambda: instance._auto_timer.stop())
-                logger.info("Auto timer created for embedded mode")
-            except ImportError as e:
-                logger.warning("EventTimer not available: %s, auto_timer disabled", e)
-                instance._auto_timer = None
-        else:
-            instance._auto_timer = None
-
-        # Register singleton
-        if singleton is not None:
-            cls._singleton_registry[singleton] = instance
-            logger.info(f"Registered singleton instance: '{singleton}'")
-
-        # Auto show (only for standalone mode, embedded mode is controlled by Qt)
-        if auto_show and not is_embedded:
-            instance.show()
-
-        return instance
-
-    @classmethod
-    def run_embedded(
-        cls,
-        title: str = "AuroraView",
-        *,
-        url: Optional[str] = None,
-        html: Optional[str] = None,
-        width: int = 800,
-        height: int = 600,
-        resizable: bool = True,
-        frame: Optional[bool] = None,
-        parent: Optional[int] = None,
-        mode: Literal["auto", "owner", "child"] = "owner",
-        bridge: Union["Bridge", bool, None] = None,  # type: ignore
-        debug: bool = True,
-        context_menu: bool = True,
-        auto_timer: bool = True,
-    ) -> "WebView":
-        """Create and show an embedded WebView with auto timer (non-blocking).
-
-        This is a convenience helper equivalent to:
-            WebView.create(..., parent=..., mode=..., auto_timer=True, auto_show=True)
-
-        Returns:
-            WebView: The created instance (kept alive by your reference)
-        """
-        instance = cls.create(
-            title=title,
-            url=url,
-            html=html,
-            width=width,
-            height=height,
-            resizable=resizable,
-            frame=frame,
-            parent=parent,
-            mode=mode,
-            bridge=bridge,
-            debug=debug,
-            context_menu=context_menu,
-            auto_show=True,
-            auto_timer=auto_timer,
-        )
-        return instance
-
-    @classmethod
-    def create_embedded(
-        cls,
-        parent_hwnd: int,
-        *,
-        title: str = "Embedded WebView",
-        width: int = 800,
-        height: int = 600,
-        url: Optional[str] = None,
-        html: Optional[str] = None,
-        asset_root: Optional[str] = None,
-        debug: bool = True,
-    ) -> "WebView":
-        """Create a WebView directly embedded into a parent window's HWND.
-
-        This is the fastest way to embed a WebView into a host application because:
-        1. No Qt Widget intermediate layer
-        2. WebView2 is created synchronously on the calling thread
-        3. Uses host's native message loop directly
-
-        This method is ideal when you have the host window's HWND and want
-        maximum performance with minimal overhead. Works with DCC applications
-        (Maya, 3ds Max, Houdini, etc.) or any Windows application with a HWND.
-
-        Args:
-            parent_hwnd: The HWND of the parent window (e.g., from Qt winId())
-            title: Window title (for debugging/identification)
-            width: Width in pixels
-            height: Height in pixels
-            url: URL to load (optional)
-            html: HTML content to load (optional)
-            asset_root: Root directory for auroraview:// protocol (optional)
-            debug: Enable developer tools (default: True)
-
-        Returns:
-            WebView: A configured WebView instance ready to use
-
-        Example (Houdini):
-            >>> import hou
-            >>> from auroraview import WebView
-            >>> from PySide2.QtCore import QTimer
-            >>>
-            >>> # Get Houdini main window HWND
-            >>> main_window = hou.qt.mainWindow()
-            >>> hwnd = int(main_window.winId())
-            >>>
-            >>> # Create WebView directly embedded
-            >>> webview = WebView.create_embedded(
-            ...     parent_hwnd=hwnd,
-            ...     title="My Tool",
-            ...     width=650,
-            ...     height=500,
-            ...     url="https://example.com"
-            ... )
-            >>>
-            >>> # Set up timer to process messages
-            >>> timer = QTimer()
-            >>> timer.timeout.connect(webview.process_events_ipc_only)
-            >>> timer.start(16)  # 60 FPS
-        """
-        from auroraview._core import WebView as _CoreWebView
-
-        logger.info(f"[create_embedded] Creating WebView for parent HWND: {parent_hwnd}")
-
-        # Create core WebView using create_embedded static method
-        core = _CoreWebView.create_embedded(
-            parent_hwnd=parent_hwnd,
-            title=title,
-            width=width,
-            height=height,
-        )
-
-        # Create Python wrapper
-        instance = cls.__new__(cls)
-        instance._core = core
-        instance._parent = parent_hwnd
-        instance._mode = "child"
-        instance._bridge = None
-        instance._auto_timer = None
-        instance._show_thread = None
-        instance._async_core = None
-        instance._async_core_lock = threading.Lock()
-        instance._close_requested = False
-        instance._event_processor = None
-
-        instance._post_eval_js_hook = None
-        instance._config = {
-            "title": title,
-            "width": width,
-            "height": height,
-            "url": url,
-            "html": html,
-            "asset_root": asset_root,
-            "debug": debug,
-        }
-
-        # Configure asset root
-        if asset_root:
-            core.set_asset_root(asset_root)
-
-        # Load content
-        if url:
-            core.load_url(url)
-        elif html:
-            core.load_html(html)
-
-        logger.info("[create_embedded] WebView created successfully")
-        logger.info("[create_embedded] Remember to call process_events_ipc_only() periodically!")
-
-        return instance
 
     def set_event_processor(self, processor: Any) -> None:
         """Set event processor (strategy pattern for UI framework integration).
@@ -1341,70 +904,6 @@ class WebView(
         """
         return self._dcc_mode
 
-    def close(self) -> None:
-        """Close the WebView window and remove from registries."""
-        logger.info("Closing WebView")
-
-        # Teardown telemetry before closing
-        self._teardown_telemetry()
-
-        # Mark close intent early so background thread can bail out if it hasn't
-        # entered the event loop yet.
-        self._close_requested = True
-
-        # Prefer closing the background-thread core (if present). Fall back to
-        # the main-thread core as a best-effort.
-        cores = []
-        try:
-            with self._async_core_lock:
-                if self._async_core is not None:
-                    cores.append(self._async_core)
-        except Exception:
-            # Lock acquisition should never fail, but keep close best-effort.
-            pass
-
-        cores.append(getattr(self, "_core", None))
-
-        seen = set()
-        for core in cores:
-            if core is None:
-                continue
-            core_id = id(core)
-            if core_id in seen:
-                continue
-            seen.add(core_id)
-
-            try:
-                core.close()
-                logger.info("Core WebView close requested")
-            except Exception as e:
-                logger.warning(f"Error requesting core close: {e}")
-
-        # Wait for background thread if running
-        if self._show_thread is not None and self._show_thread.is_alive():
-            logger.info("Waiting for background thread to finish...")
-            self._show_thread.join(timeout=5.0)
-            if self._show_thread.is_alive():
-                logger.warning("Background thread did not finish within timeout")
-            else:
-                logger.info("Background thread finished successfully")
-
-        # Remove from singleton registry
-        for key, instance in list(self._singleton_registry.items()):
-            if instance is self:
-                del self._singleton_registry[key]
-                logger.info(f"Removed from singleton registry: '{key}'")
-                break
-
-        # Remove from WindowManager
-        if self._window_id:
-            from .window_manager import get_window_manager
-
-            wm = get_window_manager()
-            wm.unregister(self._window_id)
-            logger.debug(f"WebView unregistered from WindowManager: {self._window_id}")
-
-        logger.info("WebView closed successfully")
 
     def _setup_lifecycle_events(self) -> None:
         """Setup internal event handlers for lifecycle tracking."""
@@ -1468,73 +967,4 @@ class WebView(
         """Context manager exit."""
         self.close()
 
-    # Bridge integration methods
 
-    def _setup_bridge_integration(self):
-        """Setup bidirectional communication between Bridge and WebView.
-
-        This method is called automatically when a Bridge is associated with the WebView.
-        It sets up:
-        1. Bridge → WebView: Forward bridge events to WebView UI
-        2. WebView → Bridge: Register handler to send commands to bridge clients
-        """
-        if not self._bridge:
-            return
-
-        logger.info("Setting up Bridge ↔ WebView integration")
-
-        # Bridge → WebView: Forward events
-        def bridge_callback(action: str, data: Dict, result: Any):
-            """Forward bridge events to WebView UI."""
-            logger.debug(f"Bridge event: {action}")
-            # Emit event to JavaScript with 'bridge:' prefix
-            self.emit(f"bridge:{action}", {"action": action, "data": data, "result": result})
-
-        self._bridge.set_webview_callback(bridge_callback)
-
-        # WebView → Bridge: Register command sender
-        @self.on("send_to_bridge")
-        def handle_send_to_bridge(data):
-            """Send command from WebView to Bridge clients."""
-            command = data.get("command")
-            params = data.get("params", {})
-            logger.debug(f"WebView → Bridge: {command}")
-            if self._bridge:
-                self._bridge.execute_command(command, params)
-            return {"status": "sent"}
-
-        logger.info("Bridge <-> WebView integration complete")
-
-    @property
-    def bridge(self) -> Optional["Bridge"]:  # type: ignore
-        """Get the associated Bridge instance.
-
-        Returns:
-            Bridge instance or None if no bridge is associated
-
-        Example:
-            >>> webview = WebView.create("Tool", bridge=True)
-            >>> print(webview.bridge)  # Bridge(ws://localhost:9001, ...)
-            >>>
-            >>> # Register handlers on the bridge
-            >>> @webview.bridge.on('custom_event')
-            >>> async def handle_custom(data, client):
-            ...     return {"status": "ok"}
-        """
-        return self._bridge
-
-    def send_to_bridge(self, command: str, params: Dict[str, Any] = None):
-        """Send command to Bridge clients (convenience method).
-
-        Args:
-            command: Command name
-            params: Command parameters
-
-        Example:
-            >>> webview.send_to_bridge('create_layer', {'name': 'New Layer'})
-        """
-        if not self._bridge:
-            logger.warning("No bridge associated with this WebView")
-            return
-
-        self._bridge.execute_command(command, params)
